+++ users/tests/test_artist_legal_profile_api.py
@@ -5,6 +5,7 @@ TODO валидация.
 
 import pytest
 from rest_framework import status
+from rest_framework.response import Response
 
 from users.models import (
     ArtistBankData,
@@ -16,8 +17,22 @@ from users.models import (
 pytestmark = pytest.mark.django_db
 
 
-class TestArtistLegalProfileAPI:
-    """Тесты эндпоинта юридических данных артиста."""
+@pytest.fixture
+def patch_legal_profile(artist_client, artist_legal_url):
+    """PATCH юридических данных артиста."""
+
+    def _patch(payload) -> Response:
+        return artist_client.patch(
+            artist_legal_url,
+            data=payload,
+            format='json',
+        )
+
+    return _patch
+
+
+class TestArtistLegalProfileGet:
+    """Тесты получения юридических данных артиста."""
 
     def test_artist_can_get_non_existent_legal_profile_with_fields(
         self,
@@ -28,15 +43,17 @@ class TestArtistLegalProfileAPI:
         """Артист получит не пустой объект при отсутствии юр профиля."""
         artist_user.phone = '+71234560000'
         artist_user.save(update_fields=['phone'])
+
         response = artist_client.get(artist_legal_url)
+
         assert response.status_code == status.HTTP_200_OK
         assert 'legal_profile' in response.data
         assert 'identity_data' in response.data
         assert 'bank_data' in response.data
         assert 'company_data' in response.data
-        assert artist_user.email == response.data['legal_profile']['email']
-        assert (
-            str(artist_user.phone) == response.data['legal_profile']['phone']
+        assert response.data['legal_profile']['email'] == artist_user.email
+        assert response.data['legal_profile']['phone'] == str(
+            artist_user.phone,
         )
         assert not ArtistLegalProfile.objects.filter(user=artist_user).exists()
 
@@ -49,7 +66,9 @@ class TestArtistLegalProfileAPI:
     ):
         """Артист получает существующий профиль."""
         legal_profile = artist_legal_profile_factory(user=artist_user)
+
         response = artist_client.get(artist_legal_url)
+
         assert response.status_code == status.HTTP_200_OK
         assert response.data['legal_profile']['email'] == legal_profile.email
         assert response.data['legal_profile']['recipient_type'] == (
@@ -59,10 +78,7 @@ class TestArtistLegalProfileAPI:
         assert response.data['bank_data'] is not None
         assert response.data['company_data'] is None
         assert (
-            ArtistLegalProfile.objects.get(
-                user=artist_user,
-            )
-            == legal_profile
+            ArtistLegalProfile.objects.get(user=artist_user) == legal_profile
         )
 
     def test_other_artist_cannot_get_artist_legal_profile(
@@ -75,11 +91,13 @@ class TestArtistLegalProfileAPI:
     ):
         """Другой артист не получает чужие юридические данные."""
         legal_profile = artist_legal_profile_factory(user=artist_user)
+
         response = other_artist_client.get(artist_legal_url)
+
         assert response.status_code == status.HTTP_200_OK
         assert response.data['legal_profile']['email'] != legal_profile.email
-        assert response.data['legal_profile']['email'] == (
-            other_artist_user.email
+        assert (
+            response.data['legal_profile']['email'] == other_artist_user.email
         )
         assert response.data['identity_data'] is None
         assert response.data['bank_data'] is None
@@ -92,16 +110,22 @@ class TestArtistLegalProfileAPI:
     ):
         """Не артист не может получить юр. профиль."""
         response = listener_client.get(artist_legal_url)
+
         assert response.status_code == status.HTTP_403_FORBIDDEN
 
     def test_anon_cant_get_artist_legal_profile(
         self,
-        artist_legal_url,
         api_client,
+        artist_legal_url,
     ):
         """Анон не может получить юр данные."""
         response = api_client.get(artist_legal_url)
+
         assert response.status_code == status.HTTP_401_UNAUTHORIZED
+
+
+class TestRecipientTypeList:
+    """Тесты справочника типов получателей."""
 
     def test_authenticated_user_can_get_recipient_types(
         self,
@@ -111,13 +135,13 @@ class TestArtistLegalProfileAPI:
         """Аутентифицированный пользователь может получить типы получателей."""
         response = auth_client.get(artist_recipient_type_url)
 
-        # Минимум эти не должны быть изменены.
         expected_types = [
             {'value': '', 'label': 'Не указано'},
             {'value': 'individual_entrepreneur', 'label': 'ИП'},
             {'value': 'self_employed', 'label': 'СМЗ'},
             {'value': 'legal_entity', 'label': 'Юридическое лицо'},
         ]
+
         assert response.status_code == status.HTTP_200_OK
         for item in expected_types:
             assert item in response.data
@@ -132,20 +156,20 @@ class TestArtistLegalProfileAPI:
 
         assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
+
+class TestArtistLegalProfilePatchPermissions:
+    """Тесты доступа к изменению юридических данных артиста."""
+
     def test_artist_can_patch_legal_profile(
         self,
         artist_user,
-        artist_client,
-        artist_legal_url,
         legal_profile_payload,
+        patch_legal_profile,
     ):
         """Артисту доступен patch."""
-        response = artist_client.patch(
-            artist_legal_url,
-            data=legal_profile_payload,
-            format='json',
-        )
+        response = patch_legal_profile(legal_profile_payload)
         legal_profile = ArtistLegalProfile.objects.get(user=artist_user)
+
         assert response.status_code == status.HTTP_200_OK
         assert (
             legal_profile.email
@@ -163,6 +187,7 @@ class TestArtistLegalProfileAPI:
     ):
         """Слушателю не доступно изменение юр профиля."""
         response = listener_client.patch(artist_legal_url)
+
         assert response.status_code == status.HTTP_403_FORBIDDEN
 
     def test_anon_cannot_patch_legal_profile(
@@ -172,25 +197,47 @@ class TestArtistLegalProfileAPI:
     ):
         """Анону не доступно изменение юр. профиля."""
         response = api_client.patch(artist_legal_url)
+
         assert response.status_code == status.HTTP_401_UNAUTHORIZED
+
+    def test_put_is_not_allowed(
+        self,
+        artist_client,
+        artist_legal_url,
+    ):
+        """Put запрещен."""
+        response = artist_client.put(artist_legal_url, data={}, format='json')
+
+        assert response.status_code == status.HTTP_405_METHOD_NOT_ALLOWED
+
+    def test_delete_is_not_allowed(
+        self,
+        artist_client,
+        artist_legal_url,
+    ):
+        """Delete запрещен."""
+        response = artist_client.delete(artist_legal_url)
+
+        assert response.status_code == status.HTTP_405_METHOD_NOT_ALLOWED
+
+
+class TestArtistLegalProfilePatchCreate:
+    """Тесты создания юридических данных через PATCH."""
 
     def test_patch_creates_legal_profile_if_not_exists(
         self,
         artist_user,
-        artist_client,
         legal_profile_payload,
-        artist_legal_url,
+        patch_legal_profile,
     ):
         """Создается юр профиль."""
         assert ArtistLegalProfile.objects.count() == 0
-        response = artist_client.patch(
-            artist_legal_url,
-            data=legal_profile_payload,
-            format='json',
-        )
+
+        response = patch_legal_profile(legal_profile_payload)
+        legal_profile = ArtistLegalProfile.objects.get(user=artist_user)
+
         assert response.status_code == status.HTTP_200_OK
         assert ArtistLegalProfile.objects.filter(user=artist_user).exists()
-        legal_profile = ArtistLegalProfile.objects.get(user=artist_user)
         assert (
             legal_profile.email
             == (legal_profile_payload['legal_profile']['email'])
@@ -200,33 +247,32 @@ class TestArtistLegalProfileAPI:
     def test_patch_creates_nested_identity_data_if_not_exists(
         self,
         artist_user,
-        artist_client,
         legal_profile_payload,
         identity_data_payload,
-        artist_legal_url,
+        patch_legal_profile,
     ):
         """Создаются персональные данные."""
         assert ArtistIdentityData.objects.count() == 0
-        response = artist_client.patch(
-            artist_legal_url,
-            data={**legal_profile_payload, **identity_data_payload},
-            format='json',
-        )
-        assert response.status_code == status.HTTP_200_OK
+
+        response = patch_legal_profile({
+            **legal_profile_payload,
+            **identity_data_payload,
+        })
         legal_profile = ArtistLegalProfile.objects.get(user=artist_user)
         identity_data = ArtistIdentityData.objects.get(
             legal_profile=legal_profile,
         )
+
+        assert response.status_code == status.HTTP_200_OK
         assert (
             ArtistIdentityData.objects.filter(
                 legal_profile=legal_profile,
             ).count()
             == 1
         )
-
         assert (
             identity_data.first_name
-            == (response.data['identity_data']['first_name'])
+            == response.data['identity_data']['first_name']
         )
         assert (
             identity_data.first_name
@@ -240,34 +286,30 @@ class TestArtistLegalProfileAPI:
     def test_patch_creates_nested_bank_data_if_not_exists(
         self,
         artist_user,
-        artist_client,
         legal_profile_payload,
         bank_data_payload,
-        artist_legal_url,
+        patch_legal_profile,
     ):
         """Создаются банковские данные."""
         assert ArtistBankData.objects.count() == 0
-        response = artist_client.patch(
-            artist_legal_url,
-            data={**legal_profile_payload, **bank_data_payload},
-            format='json',
-        )
-        assert response.status_code == status.HTTP_200_OK
+
+        response = patch_legal_profile({
+            **legal_profile_payload,
+            **bank_data_payload,
+        })
         legal_profile = ArtistLegalProfile.objects.get(user=artist_user)
-        bank_data = ArtistBankData.objects.get(
-            legal_profile=legal_profile,
-        )
+        bank_data = ArtistBankData.objects.get(legal_profile=legal_profile)
+
+        assert response.status_code == status.HTTP_200_OK
         assert (
             ArtistBankData.objects.filter(
                 legal_profile=legal_profile,
             ).count()
             == 1
         )
-
-        assert bank_data.bank_name == (response.data['bank_data']['bank_name'])
+        assert bank_data.bank_name == response.data['bank_data']['bank_name']
         assert (
-            bank_data.bank_name
-            == (bank_data_payload['bank_data']['bank_name'])
+            bank_data.bank_name == bank_data_payload['bank_data']['bank_name']
         )
         assert (
             bank_data.checking_account
@@ -277,63 +319,61 @@ class TestArtistLegalProfileAPI:
     def test_patch_creates_nested_company_data_if_not_exists(
         self,
         artist_user,
-        artist_client,
         legal_profile_payload,
         company_data_payload,
-        artist_legal_url,
+        patch_legal_profile,
     ):
         """Создаются юр данные."""
         assert ArtistCompanyData.objects.count() == 0
+
         legal_profile_payload['legal_profile']['recipient_type'] = (
             ArtistLegalProfile.RecipientType.LEGAL_ENTITY
         )
-        response = artist_client.patch(
-            artist_legal_url,
-            data={**legal_profile_payload, **company_data_payload},
-            format='json',
-        )
-        assert response.status_code == status.HTTP_200_OK
+
+        response = patch_legal_profile({
+            **legal_profile_payload,
+            **company_data_payload,
+        })
         legal_profile = ArtistLegalProfile.objects.get(user=artist_user)
         company_data = ArtistCompanyData.objects.get(
             legal_profile=legal_profile,
         )
+
+        assert response.status_code == status.HTTP_200_OK
         assert (
             ArtistCompanyData.objects.filter(
                 legal_profile=legal_profile,
             ).count()
             == 1
         )
-
         assert (
             company_data.company_name
-            == (response.data['company_data']['company_name'])
+            == response.data['company_data']['company_name']
         )
         assert (
             company_data.company_name
             == (company_data_payload['company_data']['company_name'])
         )
         assert (
-            company_data.ogrn == (company_data_payload['company_data']['ogrn'])
+            company_data.ogrn == company_data_payload['company_data']['ogrn']
         )
+
+
+class TestArtistLegalProfilePatchUpdate:
+    """Тесты обновления юридических данных через PATCH."""
 
     def test_patch_updates_existing_legal_profile(
         self,
         artist_user,
-        artist_client,
         artist_legal_profile_factory,
         legal_profile_payload,
-        artist_legal_url,
+        patch_legal_profile,
     ):
         """Обновляется существующий юр. профиль."""
         legal_profile = artist_legal_profile_factory(user=artist_user)
         old_id = legal_profile.id
 
-        response = artist_client.patch(
-            artist_legal_url,
-            data=legal_profile_payload,
-            format='json',
-        )
-
+        response = patch_legal_profile(legal_profile_payload)
         legal_profile.refresh_from_db()
 
         assert response.status_code == status.HTTP_200_OK
@@ -341,7 +381,7 @@ class TestArtistLegalProfileAPI:
         assert legal_profile.id == old_id
         assert (
             legal_profile.email
-            == legal_profile_payload['legal_profile']['email']
+            == (legal_profile_payload['legal_profile']['email'])
         )
         assert (
             response.data['legal_profile']['email']
@@ -351,22 +391,16 @@ class TestArtistLegalProfileAPI:
     def test_patch_updates_existing_identity_data(
         self,
         artist_user,
-        artist_client,
         artist_legal_profile_factory,
         identity_data_payload,
-        artist_legal_url,
+        patch_legal_profile,
     ):
         """Обновляются существующие персональные данные."""
         legal_profile = artist_legal_profile_factory(user=artist_user)
         identity_data = legal_profile.identity_data
         old_id = identity_data.id
 
-        response = artist_client.patch(
-            artist_legal_url,
-            data=identity_data_payload,
-            format='json',
-        )
-
+        response = patch_legal_profile(identity_data_payload)
         identity_data.refresh_from_db()
 
         assert response.status_code == status.HTTP_200_OK
@@ -389,22 +423,16 @@ class TestArtistLegalProfileAPI:
     def test_patch_updates_existing_bank_data(
         self,
         artist_user,
-        artist_client,
         artist_legal_profile_factory,
         bank_data_payload,
-        artist_legal_url,
+        patch_legal_profile,
     ):
         """Обновляются существующие банковские данные."""
         legal_profile = artist_legal_profile_factory(user=artist_user)
         bank_data = legal_profile.bank_data
         old_id = bank_data.id
 
-        response = artist_client.patch(
-            artist_legal_url,
-            data=bank_data_payload,
-            format='json',
-        )
-
+        response = patch_legal_profile(bank_data_payload)
         bank_data.refresh_from_db()
 
         assert response.status_code == status.HTTP_200_OK
@@ -416,8 +444,7 @@ class TestArtistLegalProfileAPI:
         )
         assert bank_data.id == old_id
         assert (
-            bank_data.bank_name
-            == (bank_data_payload['bank_data']['bank_name'])
+            bank_data.bank_name == bank_data_payload['bank_data']['bank_name']
         )
         assert (
             response.data['bank_data']['bank_name']
@@ -427,10 +454,9 @@ class TestArtistLegalProfileAPI:
     def test_patch_updates_existing_company_data(
         self,
         artist_user,
-        artist_client,
         artist_legal_profile_factory,
         company_data_payload,
-        artist_legal_url,
+        patch_legal_profile,
     ):
         """Обновляются существующие юр. данные."""
         legal_profile = artist_legal_profile_factory(
@@ -441,12 +467,7 @@ class TestArtistLegalProfileAPI:
         company_data = legal_profile.company_data
         old_id = company_data.id
 
-        response = artist_client.patch(
-            artist_legal_url,
-            data=company_data_payload,
-            format='json',
-        )
-
+        response = patch_legal_profile(company_data_payload)
         company_data.refresh_from_db()
 
         assert response.status_code == status.HTTP_200_OK
@@ -466,13 +487,16 @@ class TestArtistLegalProfileAPI:
             == (company_data_payload['company_data']['company_name'])
         )
 
+
+class TestArtistLegalProfilePatchRules:
+    """Тесты бизнес-правил изменения юридических данных."""
+
     def test_patch_does_not_clear_unpatched_nested_data(
         self,
         artist_user,
-        artist_client,
         artist_legal_profile_factory,
         legal_profile_payload,
-        artist_legal_url,
+        patch_legal_profile,
     ):
         """Не очищаются не переданные данные."""
         legal_profile = artist_legal_profile_factory(
@@ -487,15 +511,13 @@ class TestArtistLegalProfileAPI:
 
         legal_profile_payload['legal_profile'].pop('phone', None)
 
-        response = artist_client.patch(
-            artist_legal_url,
-            data=legal_profile_payload,
-            format='json',
-        )
+        response = patch_legal_profile(legal_profile_payload)
+
         legal_profile.refresh_from_db()
         legal_profile.identity_data.refresh_from_db()
         legal_profile.bank_data.refresh_from_db()
         legal_profile.company_data.refresh_from_db()
+
         assert response.status_code == status.HTTP_200_OK
         assert (
             response.data['legal_profile']['email']
@@ -514,23 +536,19 @@ class TestArtistLegalProfileAPI:
     def test_patch_resets_is_verified(
         self,
         artist_user,
-        artist_client,
         artist_legal_profile_factory,
-        artist_legal_url,
         legal_profile_payload,
+        patch_legal_profile,
     ):
         """Статус подтверждения сбрасывается при обновлении."""
         legal_profile = artist_legal_profile_factory(
             user=artist_user,
             is_verified=True,
         )
-        assert legal_profile.is_verified is True
-        response = artist_client.patch(
-            artist_legal_url,
-            data=legal_profile_payload,
-            format='json',
-        )
+
+        response = patch_legal_profile(legal_profile_payload)
         legal_profile.refresh_from_db()
+
         assert response.status_code == status.HTTP_200_OK
         assert (
             legal_profile.email
@@ -539,45 +557,24 @@ class TestArtistLegalProfileAPI:
         assert legal_profile.is_verified is False
         assert response.data['legal_profile']['is_verified'] is False
 
-    def test_put_is_not_allowed(
-        self,
-        artist_client,
-        artist_legal_url,
-    ):
-        """Put запрещен."""
-        response = artist_client.put(artist_legal_url, data={}, format='json')
-        assert response.status_code == status.HTTP_405_METHOD_NOT_ALLOWED
-
-    def test_delete_is_not_allowed(
-        self,
-        artist_client,
-        artist_legal_url,
-    ):
-        """Delete запрещен."""
-        response = artist_client.delete(artist_legal_url)
-        assert response.status_code == status.HTTP_405_METHOD_NOT_ALLOWED
-
     def test_readonly_fields_do_not_change(
         self,
         artist_user,
-        artist_client,
         artist_legal_profile_factory,
         legal_profile_payload,
-        artist_legal_url,
+        patch_legal_profile,
     ):
         """Поля is_verified и comment не меняются артистом."""
-        legal_profile = artist_legal_profile_factory(artist_user)
+        legal_profile = artist_legal_profile_factory(user=artist_user)
         old_is_verified = legal_profile.is_verified
         old_comment = legal_profile.comment
 
         legal_profile_payload['legal_profile']['is_verified'] = True
         legal_profile_payload['legal_profile']['comment'] = 'new comment'
-        response = artist_client.patch(
-            artist_legal_url,
-            data=legal_profile_payload,
-            format='json',
-        )
+
+        response = patch_legal_profile(legal_profile_payload)
         legal_profile.refresh_from_db()
+
         assert response.status_code == status.HTTP_200_OK
         assert response.data['legal_profile']['is_verified'] is old_is_verified
         assert legal_profile.is_verified is old_is_verified
